ingestors/massive/pyproject.toml
[project]
name = "massive"
version = "0.1.0"
description = "connect to massive.com api"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.33",
    "google-cloud-storage>=3.8.0",
]
